replaced println with asserts
