--- core-kotlin-modules/core-kotlin-7/src/main/kotlin/com/baeldung/swapfunction/SwapValue.kt
@@ -1,16 +1,0 @@
-package com.baeldung.swapfunction
-
-class SwapValue {
-
-    var a = 1
-    var b = 2
-    fun swapValues(){
-        println("Before swap: a => $a; b => $b")
-        val temp = a
-        a = b
-        b = temp
-        println("After swap: a => $a; b => $b")
-    }
-}
-
-
