--- core-kotlin-modules/core-kotlin-7/src/main/kotlin/com/baeldung/swapfunction/SwapWithAlso.kt
@@ -1,14 +1,0 @@
-package com.baeldung.swapfunction
-
-class SwapWithAlso {
-    var a = 1
-    var b = 2
-    fun swapWithAlso(){
-
-        println("Before swap: a => $a; b => $b")
-        val temp = a
-        a = b
-        b = temp
-        println("After swap: a => $a; b => $b")
-    }
-}
