+++ core-kotlin-modules/core-kotlin-7/src/test/kotlin/com/baeldung/swapfunction/classes/SwapWithPair.kt
@@ -1,4 +1,7 @@
-package com.baeldung.swapfunction
+package com.baeldung.swapfunction.classes
+
+
+import junit.framework.TestCase.assertEquals
 
 fun <T> swap(a: T, b: T): Pair<T, T> {
     return Pair(b, a)
@@ -9,10 +12,12 @@ class SwapWithPair {
     var y = 200
 
     fun swapUsingDestructuring(){
-        println("Before swap: x = $x, y = $y")
+        assertEquals(100, x)
+        assertEquals(200, y)
         val result = swap(x, y)
         x = result.first
         y = result.second
-        println("After swap: x = $x, y = $y")
+        assertEquals(200, x)
+        assertEquals(100, y)
     }
 }
